fix: build issues for rn 0.77-0.79 versions
