+++ android/local/repo/com/facetec/sdk/9.7.97/sdk-9.7.97.pom
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>com.facetec</groupId>
+  <artifactId>sdk</artifactId>
+  <version>9.7.97</version>
+  <packaging>aar</packaging>
+</project>
