+++ android/local/repo/com/facetec/sdk/maven-metadata.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<metadata>
+  <groupId>com.facetec</groupId>
+  <artifactId>sdk</artifactId>
+  <versioning>
+    <latest>9.7.97</latest>
+    <release>9.7.97</release>
+    <versions>
+      <version>9.7.97</version>
+    </versions>
+    <lastUpdated>20251127121541</lastUpdated>
+  </versioning>
+</metadata>
