+++ android/local/repo/com/kyc/ondato/nfc-reader/maven-metadata.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<metadata>
+  <groupId>com.kyc.ondato</groupId>
+  <artifactId>nfc-reader</artifactId>
+  <versioning>
+    <latest>3.2.1</latest>
+    <release>3.2.1</release>
+    <versions>
+      <version>3.2.1</version>
+    </versions>
+    <lastUpdated>20251127121558</lastUpdated>
+  </versioning>
+</metadata>
