+++ android/local/repo/com/kyc/ondato/screen-recorder/3.2.1/screen-recorder-3.2.1.module
@@ -1,0 +1,138 @@
+{
+  "formatVersion": "1.1",
+  "component": {
+    "group": "com.kyc.ondato",
+    "module": "screen-recorder",
+    "version": "3.2.1",
+    "attributes": {
+      "org.gradle.status": "release"
+    }
+  },
+  "createdBy": {
+    "gradle": {
+      "version": "8.13"
+    }
+  },
+  "variants": [
+    {
+      "name": "releaseVariantReleaseApiPublication",
+      "attributes": {
+        "org.gradle.category": "library",
+        "org.gradle.dependency.bundling": "external",
+        "org.gradle.libraryelements": "aar",
+        "org.gradle.usage": "java-api"
+      },
+      "dependencies": [
+        {
+          "group": "androidx.databinding",
+          "module": "viewbinding",
+          "version": {
+            "requires": "8.11.1"
+          }
+        },
+        {
+          "group": "org.jetbrains.kotlin",
+          "module": "kotlin-stdlib",
+          "version": {
+            "requires": "2.0.21"
+          }
+        }
+      ],
+      "files": [
+        {
+          "name": "screen-recorder-3.2.1.aar",
+          "url": "screen-recorder-3.2.1.aar",
+          "size": 70687,
+          "sha512": "f75c041870c06db04333601ade95f856e43c0cdb927cb2eff58d274fe4b9825fb98bb30af87f4b57317b3d492dcabfdda889a276ed02ed6c2a0a819503ee4121",
+          "sha256": "b51d78b56bcd48c8762992ebbe01662fd72bc44ae3b0405b4aae294b5889531e",
+          "sha1": "9ac303b482188a948978e83c06bdebd9869f1f7b",
+          "md5": "bcdb335c486b1573afbcca9b20f46d34"
+        }
+      ]
+    },
+    {
+      "name": "releaseVariantReleaseRuntimePublication",
+      "attributes": {
+        "org.gradle.category": "library",
+        "org.gradle.dependency.bundling": "external",
+        "org.gradle.libraryelements": "aar",
+        "org.gradle.usage": "java-runtime"
+      },
+      "dependencies": [
+        {
+          "group": "com.kyc.ondato",
+          "module": "sdk-core",
+          "version": {
+            "requires": "3.2.1"
+          }
+        },
+        {
+          "group": "androidx.core",
+          "module": "core-ktx",
+          "version": {
+            "requires": "1.16.0"
+          }
+        },
+        {
+          "group": "androidx.appcompat",
+          "module": "appcompat",
+          "version": {
+            "requires": "1.7.1"
+          }
+        },
+        {
+          "group": "com.facetec",
+          "module": "sdk",
+          "version": {
+            "requires": "9.7.97"
+          },
+          "excludes": [
+            {
+              "group": "*",
+              "module": "*"
+            }
+          ],
+          "thirdPartyCompatibility": {
+            "artifactSelector": {
+              "name": "sdk",
+              "type": "aar",
+              "extension": "aar"
+            }
+          }
+        },
+        {
+          "group": "androidx.localbroadcastmanager",
+          "module": "localbroadcastmanager",
+          "version": {
+            "requires": "1.1.0"
+          }
+        },
+        {
+          "group": "androidx.databinding",
+          "module": "viewbinding",
+          "version": {
+            "requires": "8.11.1"
+          }
+        },
+        {
+          "group": "org.jetbrains.kotlin",
+          "module": "kotlin-stdlib",
+          "version": {
+            "requires": "2.0.21"
+          }
+        }
+      ],
+      "files": [
+        {
+          "name": "screen-recorder-3.2.1.aar",
+          "url": "screen-recorder-3.2.1.aar",
+          "size": 70687,
+          "sha512": "f75c041870c06db04333601ade95f856e43c0cdb927cb2eff58d274fe4b9825fb98bb30af87f4b57317b3d492dcabfdda889a276ed02ed6c2a0a819503ee4121",
+          "sha256": "b51d78b56bcd48c8762992ebbe01662fd72bc44ae3b0405b4aae294b5889531e",
+          "sha1": "9ac303b482188a948978e83c06bdebd9869f1f7b",
+          "md5": "bcdb335c486b1573afbcca9b20f46d34"
+        }
+      ]
+    }
+  ]
+}
