+++ android/local/repo/com/kyc/ondato/sdk-core/maven-metadata.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<metadata>
+  <groupId>com.kyc.ondato</groupId>
+  <artifactId>sdk-core</artifactId>
+  <versioning>
+    <latest>3.2.1</latest>
+    <release>3.2.1</release>
+    <versions>
+      <version>3.2.1</version>
+    </versions>
+    <lastUpdated>20251127121543</lastUpdated>
+  </versioning>
+</metadata>
