+++ example/android/app/build.gradle
@@ -110,10 +110,10 @@ android {
 dependencies {
   // The version of react-native is set by the React Native Gradle Plugin
   implementation("com.facebook.react:react-android")
-  implementation("com.kyc.ondato:screen-recorder:3.2.0") {
+  implementation("com.kyc.ondato:screen-recorder:3.2.1") {
     exclude group: "com.squareup.okhttp3"
   }
-  implementation("com.kyc.ondato:nfc-reader:3.2.0") {
+  implementation("com.kyc.ondato:nfc-reader:3.2.1") {
     exclude group: "com.squareup.okhttp3"
   }
