+++ example/ios/OndatoSdkReactNativeExample/Images.xcassets/AppIcon.appiconset/Contents.json
@@ -1,53 +1,53 @@
 {
-  "images": [
+  "images" : [
     {
-      "idiom": "iphone",
-      "scale": "2x",
-      "size": "20x20"
+      "idiom" : "iphone",
+      "scale" : "2x",
+      "size" : "20x20"
     },
     {
-      "idiom": "iphone",
-      "scale": "3x",
-      "size": "20x20"
+      "idiom" : "iphone",
+      "scale" : "3x",
+      "size" : "20x20"
     },
     {
-      "idiom": "iphone",
-      "scale": "2x",
-      "size": "29x29"
+      "idiom" : "iphone",
+      "scale" : "2x",
+      "size" : "29x29"
     },
     {
-      "idiom": "iphone",
-      "scale": "3x",
-      "size": "29x29"
+      "idiom" : "iphone",
+      "scale" : "3x",
+      "size" : "29x29"
     },
     {
-      "idiom": "iphone",
-      "scale": "2x",
-      "size": "40x40"
+      "idiom" : "iphone",
+      "scale" : "2x",
+      "size" : "40x40"
     },
     {
-      "idiom": "iphone",
-      "scale": "3x",
-      "size": "40x40"
+      "idiom" : "iphone",
+      "scale" : "3x",
+      "size" : "40x40"
     },
     {
-      "idiom": "iphone",
-      "scale": "2x",
-      "size": "60x60"
+      "idiom" : "iphone",
+      "scale" : "2x",
+      "size" : "60x60"
     },
     {
-      "idiom": "iphone",
-      "scale": "3x",
-      "size": "60x60"
+      "idiom" : "iphone",
+      "scale" : "3x",
+      "size" : "60x60"
     },
     {
-      "idiom": "ios-marketing",
-      "scale": "1x",
-      "size": "1024x1024"
+      "idiom" : "ios-marketing",
+      "scale" : "1x",
+      "size" : "1024x1024"
     }
   ],
-  "info": {
-    "author": "xcode",
-    "version": 1
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
   }
 }
