--- ios/OSRNModule.h
@@ -1,5 +1,0 @@
-#import <OndatoSpec/OndatoSpec.h>
-
-@interface OSRNModule : NSObject <NativeOndatoModuleSpec>
-
-@end
