+++ ios/OndatoModule.h
@@ -1,0 +1,5 @@
+#import <RNOndatoSpec/RNOndatoSpec.h>
+
+@interface OndatoModule : NSObject <NativeOndatoModuleSpec>
+
+@end
