+++ plugin/src/constants.ts
@@ -1,7 +1,7 @@
 export const MIN_SDK_VERSION = 24;
 export const DEPLOYMENT_TARGET = '15.1';
 
-export const ONDATO_VERSION_ANDROID = '3.2.0';
+export const ONDATO_VERSION_ANDROID = '3.2.1';
 export const ONDATO_VERSION_IOS = '3.2.0';
 
 export const MAVEN_REPOS = [
